OSINT Tools
Objective: Determine all working links without compromising existing organized structure. Identify country where each site is hosted e.g. USA, Israel, Russia etc.

Details: There are 5124 links and of those links I separated the domain out and tested them all. Of the list - 4286 responded, with a 200 and the remaining 838 responded with some form of error. 18 of the 838 errors are because the link is an improper URL (e.g. ".." is in the middle of the URL indicating that text was copied and not the link). Of the total number there are also an unknown number of links that are (as of last week) no longer valid because the product linked is no longer hosted at that location.